# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Coupang AI Detail Maker API
  version: 0.1.0
  description: |
    Coupang AI Detail Maker 백엔드 REST API 명세.

    - 기준 문서: PRD v0.3.9(8장 주요 API가 엔드포인트 원천), 도메인 정의서 v0.3.10, 구조 원칙 v0.1.10, ERD v0.1.10
    - 문서가 갱신되면 이 명세도 갱신한다.
    - JSON 필드는 camelCase(NM-10). 오류는 `{ "error": { "code", "message" } }` 하나로 통일(NM-11), code는 구조 원칙 4.2 표만 쓴다.
    - 판정 우선순위(FR-06, 구조 원칙 4.2): 401 → 404 → 409(PUBLISHED 대상) → 403(이메일 미인증) → 402(잔액 부족) → 409(version·진행 중 작업) → 429 → 503. 미인증과 잔액 0이 겹치면 403.
    - 레이트 리밋(NFR-04): 일반 API 사용자당 60 req/min(인증 뒤에 적용), 로그인 IP당 10 req/min, 토큰 갱신 IP당 30 req/min. 초과 시 429 `RATE_LIMITED`. 가입·로그아웃은 리밋 없음.
    - uuid 형식이 아닌 프로젝트 id는 404 `NOT_FOUND`.
    - 처리되지 않은 오류는 500 `INTERNAL`(message에 내부 정보 없음). 각 operation에는 따로 적지 않는다.
    - 퍼블리시 전 모든 응답에 원본·프리뷰 키, draftHtml, finalHtml, 공개 URL이 없다(PP-05, NFR-08).
    - 표기: 우선순위 M/S, 단계 P1(2일 핵심 슬라이스)·P2(MVP 완성). S는 MVP 이후(M4).
servers:
  - url: /
    description: 단일 도메인·동일 출처(DEC-02). Express가 frontend/dist와 /api를 함께 서빙한다.
...
security:
  - bearerAuth: []
paths:
...
  /api/projects/{id}/publish:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [publish]
      operationId: publishProject
      summary: 퍼블리시 TX(크레딧 1 차감) `*`
      description: |
        M·P1(BE-14a) / P2(공개 이미지 사본, BE-14b), FR-06, FR-21, FR-22, FR-34, FR-35, BR-10~15·BR-39·BR-48·BR-52·BR-66, P-5.
        한 TX: 프로젝트 행 `FOR UPDATE` → 이미 PUBLISHED면 version·자격 검사 없이 기존 결과 200(BR-13, 멱등, 추가 차감 없음) →
        이메일 인증(403)·잔액(402) 선검사 → version·진행 중 작업(409) → DEDUCT 원장(프로젝트당 1건, `ON CONFLICT DO NOTHING`) → 잔액 −1 →
        finalHtml 저장·PUBLISHED·version+1 → PublishRecord(PENDING). 하나라도 실패하면 ROLLBACK(잔액·상태 불변).
        잔액 0이면 version이 틀려도 402. 동시 2건은 둘 다 200, DEDUCT 1건, finalHtml 동일.
        COMMIT 뒤 원본을 만료 없는 공개 URL로 복사한다(`{assetId}.{ext}`, 최종 HTML img는 `PUBLIC_IMAGE_BASE_URL/{assetId}.{ext}`). 복사는 최대 3회 시도하고 실패하면 로그만 남기며 크레딧은 유지한다.
        실패한 사본은 퍼블리시 재요청(PUBLISHED 포함)과 `GET /final`이 응답 전에 재시도하고, 응답은 사본 준비 여부와 무관하게 최종 HTML이다(I-16 해소).
        판정 순서: 404 → PUBLISHED(기존 결과 200) → 403 → 402 → 409 version·진행 중 작업·상태(GENERATED·EDITING이 아니면 `INVALID_STATE`) → 429.
      requestBody:
        $ref: '#/components/requestBodies/VersionOnly'
      responses:
        '200':
          description: 퍼블리시 완료 또는 기존 결과, 최종 HTML
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalHtml'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/projects/{id}/final:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [publish]
      operationId: getFinalHtml
      summary: 최종 HTML 조회(PUBLISHED만)
      description: |
        M·P1(BE-14a, 확장 토큰 허용은 S BE-20), FR-22, FR-23, BR-52·BR-60·BR-62. PUBLISHED가 아니면 403 `NOT_PUBLISHED`.
        자격 무관(잔액 0이어도 200), 재조회 무제한, 잔액 불변. Access Token 또는 해당 프로젝트 범위 확장 토큰(`typ=ext`)으로 호출.
        응답 전에 아직 복사되지 않은 공개 이미지 사본을 다시 시도한다(I-16).
      security:
        - bearerAuth: []
        - extensionToken: []
      responses:
        '200':
          description: 최종 HTML
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalHtml'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access Token(HS256, `typ=access`, 15분). 로그인·갱신 응답 body로 받아 메모리에만 보관. 만료 401 `TOKEN_EXPIRED`, 서명·`typ` 오류 401 `TOKEN_INVALID`.
    extensionToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: (S) 확장 토큰(`typ=ext`, `prj`=projectId, 10분). `GET /final`, `POST /publish-report`에서만 허용.
    refreshCookie:
      type: apiKey
      in: cookie
      name: rt
      description: Refresh Token(HttpOnly·Secure·SameSite=Strict·Path=/api/auth, 14일, 회전). `/api/auth/refresh`·`/logout`은 Origin도 검사.
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      description: 프로젝트 ID. 남의 프로젝트는 404
      schema:
        type: string
        format: uuid
...
  requestBodies:
    VersionOnly:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [version]
            properties:
              version:
                $ref: '#/components/schemas/Version'
  responses:
    BadRequest:
      description: 입력 검증 실패 `VALIDATION_FAILED`(편집 text에 태그, URL 패턴 외, 업로드 제한, 필수값 누락, 중복 이메일 가입, USP 0개 저장 등)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_FAILED
              message: 입력값을 확인하세요
    Unauthorized:
      description: 인증 실패. `TOKEN_EXPIRED`(만료, 1회 갱신 후 재시도) / `TOKEN_INVALID`(서명·`typ` 오류, 갱신 없이 로그인) / `INVALID_CREDENTIALS`(로그인 실패) / `REFRESH_INVALID`(refresh 무효·재사용·폐기)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: TOKEN_EXPIRED
              message: 토큰이 만료되었습니다
    PaymentRequired:
      description: 잔액 부족 `INSUFFICIENT_CREDIT`
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: '`EMAIL_NOT_VERIFIED`(이메일 미인증, 잔액 0과 겹쳐도 403 우선) / `ORIGIN_FORBIDDEN`(Origin 불일치) / `NOT_PUBLISHED`(미퍼블리시 final 조회)'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: '`NOT_FOUND`. 없는 리소스, 남의 프로젝트(존재 비노출), uuid 형식이 아닌 프로젝트 id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: '`INVALID_STATE`(허용되지 않는 상태: PUBLISHED 변경 요청, 생성 전 preview, DRAFT·ANALYZED 퍼블리시 등) / `VERSION_CONFLICT`(version 불일치) / `JOB_IN_PROGRESS`(진행 중 LLM 작업). 클라이언트는 프로젝트를 다시 조회'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VERSION_CONFLICT
              message: 다른 곳에서 먼저 변경되었습니다
    TooManyRequests:
      description: '한도 초과, LLM 미호출. `ANALYZE_LIMIT` / `REGEN_LIMIT` / `AI_EDIT_LIMIT` / `DAILY_LLM_LIMIT`(Asia/Seoul 자정 초기화, 초기화 시각 안내 필드는 확인 필요) / `RATE_LIMITED`(NFR-04)'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: 구조 원칙 4.2 오류 코드
              enum:
                - VALIDATION_FAILED
                - TOKEN_EXPIRED
                - TOKEN_INVALID
                - INVALID_CREDENTIALS
                - REFRESH_INVALID
                - INSUFFICIENT_CREDIT
                - EMAIL_NOT_VERIFIED
                - ORIGIN_FORBIDDEN
                - NOT_PUBLISHED
                - NOT_FOUND
                - VERSION_CONFLICT
                - JOB_IN_PROGRESS
                - INVALID_STATE
                - ANALYZE_LIMIT
                - REGEN_LIMIT
                - AI_EDIT_LIMIT
                - DAILY_LLM_LIMIT
                - RATE_LIMITED
                - UPSTREAM_FAILED
                - LLM_BUSY
                - INTERNAL
            message:
              type: string
              description: 사용자 안내 문구. 내부 정보 노출 금지
    Version:
      type: integer
      minimum: 1
      description: 기준 프로젝트 version(낙관적 잠금, FR-34). 현재 값과 다르면 409
...
    FinalHtml:
      type: object
      description: 최종 HTML(PUBLISHED만). 워터마크·`data-edit-id`·`data-block-id` 제거, 인라인 CSS, 만료 없는 공개 이미지 URL
      properties:
        finalHtml:
          type: string
...